requirements.txt
requests
pyyaml
sqlalchemy
pandas
duckdb
duckdb-engine
beautifulsoup4
pytz